Create Main.java
The Levels one to ten are working fine, and there is a fade between level ten and eleven, but the collision on level eleven seems to be all messed up, but I have checked through the code multiple times, and I can not see the problem. Level ten has no gravity, and level eleven's gravity does not work, even though I turned it back on, I know there are a few other bugs, like sometimes you respawn in the wrong place, and you can clip into the platforms a bit, but I'm mostly worried about why the gravity and collision on level eleven is not working, there are also no errors, the collision just doesn't work. If someone could tell me why this is that would be great, thankyou!